Revert "Merge pull request #152 from ClimateMind/CM-329_bug_fix_subscribe_code"
This reverts commit 221280580948d16d1e07423c8a89767e49145a7a, reversing
changes made to 4768aa43918a62984ebfa73a9302c0638ac79e61.

I accidentally pushed to develop and this should undo that.
