+++ migrations_local/versions/1d9287012974_add_all_current_tables.py
@@ -1,8 +1,8 @@
 """add all current tables
 
-Revision ID: 2738cace3ef7
+Revision ID: 1d9287012974
 Revises: 
-Create Date: 2021-01-13 18:56:32.376009
+Create Date: 2021-01-11 08:25:21.410084
 
 """
 from alembic import op
@@ -10,7 +10,7 @@ import sqlalchemy as sa
 
 
 # revision identifiers, used by Alembic.
-revision = "2738cace3ef7"
+revision = "1d9287012974"
 down_revision = None
 branch_labels = None
 depends_on = None
@@ -24,6 +24,12 @@ def upgrade():
         sa.Column("postal_code", sa.String(length=5), nullable=True),
         sa.Column("ip_address", sa.String(length=255), nullable=True),
         sa.PrimaryKeyConstraint("session_id"),
+    )
+    op.create_table(
+        "signup",
+        sa.Column("email", sa.String(length=254), nullable=False),
+        sa.Column("signup_timestamp", sa.DateTime(), nullable=True),
+        sa.PrimaryKeyConstraint("email"),
     )
     op.create_table(
         "users",
@@ -79,27 +85,16 @@ def upgrade():
         ),
         sa.PrimaryKeyConstraint("scores_id"),
     )
-    op.create_table(
-        "signup",
-        sa.Column("email", sa.String(length=254), nullable=False),
-        sa.Column("session_id", sa.String(length=256), nullable=True),
-        sa.Column("signup_timestamp", sa.DateTime(), nullable=True),
-        sa.ForeignKeyConstraint(
-            ["session_id"],
-            ["sessions.session_id"],
-        ),
-        sa.PrimaryKeyConstraint("email"),
-    )
     # ### end Alembic commands ###
 
 
 def downgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table("signup")
     op.drop_table("scores")
     op.drop_table("climate_feed")
     op.drop_index(op.f("ix_users_username"), table_name="users")
     op.drop_index(op.f("ix_users_email"), table_name="users")
     op.drop_table("users")
+    op.drop_table("signup")
     op.drop_table("sessions")
     # ### end Alembic commands ###
